refactor(payment): CHECKOUT-3011 Move `Wepay` files to payment folder
