+++ src/payment/strategies/index.ts
@@ -7,10 +7,10 @@ export { default as PaymentStrategy } from './payment-strategy';
 export { default as PaypalExpressPaymentStrategy } from './paypal-express-payment-strategy';
 export { default as PaypalProPaymentStrategy } from './paypal-pro-payment-strategy';
 export { default as SagePayPaymentStrategy } from './sage-pay-payment-strategy';
-export { default as WepayPaymentStrategy } from './wepay-payment-strategy';
 
 export { AfterpayPaymentStrategy } from './afterpay';
 export { AmazonPayPaymentStrategy, AmazonPayPaymentInitializeOptions } from './amazon-pay';
 export { BraintreeCreditCardPaymentStrategy, BraintreePaymentInitializeOptions, BraintreePaypalPaymentStrategy, BraintreeVisaCheckoutPaymentStrategy, BraintreeVisaCheckoutPaymentInitializeOptions } from './braintree';
 export { KlarnaPaymentStrategy, KlarnaPaymentInitializeOptions } from './klarna';
 export { SquarePaymentStrategy, SquarePaymentInitializeOptions } from './square';
+export { WepayPaymentStrategy } from './wepay';
