+++ src/payment/strategies/wepay/index.ts
@@ -1,0 +1,2 @@
+export { default as WepayPaymentStrategy } from './wepay-payment-strategy';
+export { default as WepayRiskClient } from './wepay-risk-client';
