+++ src/payment/strategies/wepay/wepay-payment-strategy.ts
@@ -1,12 +1,12 @@
 import { merge } from 'lodash';
 
-import { CheckoutStore, InternalCheckoutSelectors } from '../../checkout';
-import { OrderActionCreator, OrderRequestBody } from '../../order';
-import { WepayRiskClient } from '../../remote-checkout/methods/wepay';
-import PaymentActionCreator from '../payment-action-creator';
-import { PaymentInitializeOptions, PaymentRequestOptions } from '../payment-request-options';
+import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
+import { OrderActionCreator, OrderRequestBody } from '../../../order';
+import PaymentActionCreator from '../../payment-action-creator';
+import { PaymentInitializeOptions, PaymentRequestOptions } from '../../payment-request-options';
+import CreditCardPaymentStrategy from '../credit-card-payment-strategy';
 
-import CreditCardPaymentStrategy from './credit-card-payment-strategy';
+import WepayRiskClient from './wepay-risk-client';
 
 export default class WepayPaymentStrategy extends CreditCardPaymentStrategy {
     constructor(
