--- src/remote-checkout/methods/wepay/index.ts
@@ -1,1 +1,0 @@
-export { default as WepayRiskClient } from './wepay-risk-client';
