+++ src/remote-checkout/remote-checkout-reducer.ts
@@ -1,6 +1,7 @@
 import { combineReducers, Action } from '@bigcommerce/data-store';
 
 import { AfterpayRemoteCheckout, AmazonPayRemoteCheckout } from './methods';
+
 import * as actionTypes from './remote-checkout-action-types';
 import RemoteCheckoutState, { RemoteCheckoutStateData } from './remote-checkout-state';
